+++ NEWS.md
@@ -3,7 +3,7 @@
 ## New features
 
 - `arr()` and `nnt()` for Absolute Risk Reduction or Number Needed to Treat.
-
+- `oddsratio_to_arr()`, `riskratio_to_arr()`, `nnt_to_arr()` and their inverses.
 
 # effectsize 0.8.3